pyproject.toml
[project]
name = "aerpawlib"
version = "1.0.0beta.8"
description = "Tools and frameworks for writing AERPAW scripts"
readme = "README.md"
authors = [
    {name = "John Kesler", email = "jckesle2@ncsu.edu"},
    {name = "Julian Reder", email = "jjreder@ncsu.edu"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["aerpaw", "drone", "uav", "mavsdk", "vehicle", "control"]

dependencies = [
    # Core dependencies (used by all versions)
    "pyzmq>=25.0.0",
    "pykml>=0.2.0",
    "shapely>=2.0.0",
    # V2 dependencies (MAVSDK-based, default)
    "mavsdk>=2.0.0",
    "aiohttp>=3.12.14",
    "dronekit>=2.9.2",
    "pymavlink>=2.4.0",
    "pyserial>=3.5",
    "requests>=2.32.4",
    "pyyaml"

]

[project.optional-dependencies]
# Legacy/V1 dependencies (DroneKit-based)
legacy = [
    "dronekit>=2.9.2",
    "pymavlink>=2.4.0",
    "pyserial>=3.5",
    "requests>=2.32.4",
]

dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
]
all = [
    "aerpawlib[legacy,v2,dev,docs]",
]

[project.urls]
Homepage = "https://github.com/quantumbagel/aerpawlib"
Repository = "https://github.com/quantumbagel/aerpawlib"
Documentation = "https://github.com/quantumbagel/aerpawlib#readme"
"Bug Tracker" = "https://github.com/quantumbagel/aerpawlib/issues"

[project.scripts]
aerpawlib = "aerpawlib.__main__:main"

[tool.setuptools.packages.find]
include = ["aerpawlib*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.kml", "*.plan"]

[tool.black]
line-length = 79
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["aerpawlib/test"]
python_files = ["*_test.py", "test_*.py"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["aerpawlib"]
omit = ["*/test/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

